refactor(mmap): extract into top level module
