+++ common/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod cache;
 pub mod compression;
 pub mod ipc;
+pub mod mmap;
